requirements.txt
# Core
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0

# Data
alpaca-py>=0.35.0          # Alpaca Markets REST + stream
python-binance>=1.0.19     # Binance REST + websocket
feedparser>=6.0.11         # RSS news ingestion
httpx>=0.27.0              # Async HTTP
websockets>=13.0

# Quant
pandas>=2.2.0
numpy>=2.1.0
ta>=0.11.0                 # Technical indicators (RSI, ATR, MACD…)
web3>=7.0.0                # On-chain data (Ethereum)

# Execution
nautilus_trader>=1.205.0   # Stock execution framework
hummingbot>=2.0.0; sys_platform != "darwin"  # Crypto execution (Linux only)

# Monitoring
prometheus-client>=0.21.0
schedule>=1.2.0

# Dev
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
